requirements-dev.txt
eth-brownie==1.17.0
vyper==0.3.0